API/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
starlette==0.49.3
python-multipart==0.0.20

# Core dependencies
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0

torch==2.7.1
torchvision==0.22.1
# YOLOv8 (Ultralytics)
#torch==2.7.1+cu118
#torchvision==0.22.1+cu118
ultralytics==8.3.223
ultralytics-thop==2.0.18

# Utils
requests==2.32.5
PyYAML==6.0.3
typing_extensions==4.15.0
